code/CR_recon/requirements.txt
torch>=2.0.0
numpy>=1.20.0
pyyaml>=5.4.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
websockets>=10.0
tqdm>=4.60.0
